requirements.txt
fastapi==0.110.2
uvicorn==0.30.0
pymongo==4.6.1
python-jose==3.3.0
passlib==1.7.4
requests==2.31.0
python-multipart==0.0.9
pydantic~=2.7.3